# `MemPostings.Delete()`: reduce locking/unlocking