--- a/tsdb/index/postings_test.go
+++ b/tsdb/index/postings_test.go
@@ -1002,6 +1002,8 @@ func TestMemPostings_Delete(t *testing.T) {
 	require.Empty(t, expanded, "expected empty postings, got %v", expanded)
 }
 
+// BenchmarkMemPostings_Delete is quite heavy, so consider running it with
+// -benchtime=10x or similar to get more stable and comparable results.
 func BenchmarkMemPostings_Delete(b *testing.B) {
 	internedItoa := map[int]string{}
 	var mtx sync.RWMutex