(The description has been updated since I created this PR)

`MemPostings.Delete` is called from Head.gc(), i.e. it gets the IDs of the series that have churned.

With current implementation, it tried to acquire the `Lock()` for each label/value combination: this can be quite expensive for tenants who have lots of those combinations.

This changes the implementation to call `Lock` only once per each label name (not value), and only if it is affected by the series churn.

I'm also adding a benchmark, which I ran with `-benchtime=10x`: it's quite difficult to get meaningful result here as each time the operation called, the benchmarked entity is being changed. 

The benchmark runs some read operations which would make the `Lock()` operation wait, which is the scenario that I'm trying to mitigate, as in a downstream project we've seen some profiles waiting continuously on this `Lock()` call.

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb/index
                                            │     main     │                 new                 │
                                            │    sec/op    │    sec/op     vs base               │
MemPostings_Delete/refs=1/readers=0-12         141.2m ± 5%   157.2m ±  1%  +11.29% (p=0.002 n=6)
MemPostings_Delete/refs=1/readers=1-12        1152.1m ± 1%   205.7m ±  5%  -82.14% (p=0.002 n=6)
MemPostings_Delete/refs=1/readers=10-12        10.424 ± 3%    1.149 ± 45%  -88.98% (p=0.002 n=6)
MemPostings_Delete/refs=100/readers=0-12       173.8m ± 5%   202.2m ± 13%  +16.36% (p=0.004 n=6)
MemPostings_Delete/refs=100/readers=1-12      1203.4m ± 1%   226.8m ±  5%  -81.15% (p=0.002 n=6)
MemPostings_Delete/refs=100/readers=10-12      10.791 ± 2%    1.164 ± 49%  -89.21% (p=0.002 n=6)
MemPostings_Delete/refs=10000/readers=0-12     198.3m ± 7%   176.3m ±  2%  -11.08% (p=0.002 n=6)
MemPostings_Delete/refs=10000/readers=1-12    1246.2m ± 2%   241.1m ±  4%  -80.66% (p=0.002 n=6)
MemPostings_Delete/refs=10000/readers=10-12    10.571 ± 3%    1.122 ± 49%  -89.39% (p=0.002 n=6)
geomean                                         1.292        357.2m        -72.34%
```

I haven't found the reason for such a high variance in some of the new implementation's results.

I only ran memory benchmarks with `readers=0`, and the results are very similar:
```
                                           │   main_mem    │                new_mem                │
                                           │     B/op      │     B/op       vs base                │
MemPostings_Delete/refs=1/readers=0-12       28.17Mi ± ∞ ¹   27.65Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
MemPostings_Delete/refs=100/readers=0-12     31.22Mi ± ∞ ¹   30.70Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
MemPostings_Delete/refs=10000/readers=0-12   30.37Mi ± ∞ ¹   29.84Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                      29.89Mi         29.37Mi        -1.74%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                           │   main_mem   │                new_mem                │
                                           │  allocs/op   │  allocs/op    vs base                 │
MemPostings_Delete/refs=1/readers=0-12       36.000 ± ∞ ¹    8.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
MemPostings_Delete/refs=100/readers=0-12      248.0 ± ∞ ¹    221.0 ± ∞ ¹        ~ (p=1.000 n=1) ²
MemPostings_Delete/refs=10000/readers=0-12   20.18k ± ∞ ¹   20.15k ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                                       564.7          329.0        -41.74%
```