@colega 2024-06-06
I've recently seen a head GC taking 3h on a downstream project, and it looked like a possible lock contention, so I thought it would be interesting to recover this. 

I have no idea how I could benchmark it. It looks like it only has impact under pressure.