@colega 2024-06-09
I've optimized the code further, now locking at most once per label name in the storage, pushed a benchmark and updated the PR description.