@colega 2024-06-09
FWIW, I ran the benchmark with 1 iteration removing 100k series with 100 concurrent readers: it too 265s on `main` and 1.7s on this new implementation.