@colega 2024-06-10
> I wonder if the whole operation might be more efficient if we started with the series to be deleted, and collected together each name/value pair affected. I guess it depends on the
relative size of the Postings index and the set of series to be deleted.
But that would be another PR entirely.

Right, I think we can just hook to the call made to `seriesLifecycleCallback`:

https://github.com/prometheus/prometheus/blob/08621bebe976c683683aa8198303bb000381be57/tsdb/head.go#L1966-L1966

It shouldn't be an issue to split this `Delete` call into multiple calls.